Revert "refactor(autopilots): fold access management into the edit dialog (MU…"
This reverts commit 48f49d8abc5898f4f94275e658fd8996fc21f141.
