+++ packages/views/autopilots/components/autopilot-dialog.tsx
@@ -51,7 +51,6 @@ import { buildAutopilotWebhookUrl } from "@multica/core/autopilots";
 import { api } from "@multica/core/api";
 import type {
   AutopilotAssigneeType,
-  AutopilotCollaborator,
   AutopilotExecutionMode,
   AutopilotTrigger,
 } from "@multica/core/types";
@@ -61,7 +60,6 @@ import { ProjectPicker } from "../../projects/components/project-picker";
 import { ProjectIcon } from "../../projects/components/project-icon";
 import { AgentPicker, type AssigneeSelection } from "./pickers/agent-picker";
 import { SubscriberMultiSelect } from "./subscriber-multi-select";
-import { AutopilotAccessManager } from "./autopilot-access-manager";
 import {
   getDefaultTriggerConfig,
   getLocalTimezone,
@@ -104,8 +102,6 @@ export type AutopilotDialogProps =
       autopilotId: string;
       initial: AutopilotInitial;
       triggers: AutopilotTrigger[];
-      collaborators: AutopilotCollaborator[];
-      canManageAccess: boolean;
     };
 
 // ---------------------------------------------------------------------------
@@ -669,13 +665,6 @@ export function AutopilotDialog(props: AutopilotDialogProps) {
               />
             )}
 
-            {!isCreate && props.canManageAccess && (
-              <AccessSection
-                autopilotId={props.autopilotId}
-                collaborators={props.collaborators}
-              />
-            )}
-
             {isCreate && (
               <TriggerKindSection kind={triggerKind} onChange={setTriggerKind} />
             )}
@@ -918,28 +907,6 @@ function SubscribersSection({
       <SubscriberMultiSelect
         selectedIds={selectedUserIds}
         onChange={onChange}
-      />
-    </div>
-  );
-}
-
-function AccessSection({
-  autopilotId,
-  collaborators,
-}: {
-  autopilotId: string;
-  collaborators: AutopilotCollaborator[];
-}) {
-  const { t } = useT("autopilots");
-  return (
-    <div>
-      <SectionLabel>{t(($) => $.dialog.section_access)}</SectionLabel>
-      <p className="mb-2 text-[11px] text-muted-foreground">
-        {t(($) => $.access.description)}
-      </p>
-      <AutopilotAccessManager
-        autopilotId={autopilotId}
-        collaborators={collaborators}
       />
     </div>
   );
